frontend: Refactor YoutubeApiWrappers
Created YoutubeApiClient and moved the broadcast and stream related
functions from YoutubeApiWrappers to YoutubeApiClient and added
apiClient object to YoutubeApiWrappers. This helps move the mixed
ownership issue from YoutubeApiWrappers inheriting from YoutubeAuth only
because it needs OAuth information, follow up refactor would be to move
the broadcast setup/start/stop out to YoutubeBroadcastController and
create a YoutubeService class which owns the OAuthTokenStore,
YoutubeApiClient, YoutubeBroadcastController and the
YoutubeDockController.
